Fix missing Code and Refactor Hooks and use Custom Hooks R #17
