+++ src/components/NavBar/NavBar.jsx
@@ -1,18 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
+import useNavbarHook from '../../hooks/useNavBarHook';
 
 import './navbar.css';
 
-const NavBar = () => {
-  const [height, setHeight] = useState(0);
-  window.addEventListener('scroll', () => setHeight(window.scrollY));
-  let { pathname } = useLocation();
-
-  const [path, setPath] = useState(false);
-
-  useEffect(() => {
-    setPath(pathname.includes('launch'));
-  }, [pathname]);
+export default function NavBar() {
+  const { pathname } = useLocation();
+  const [height, path] = useNavbarHook(pathname);
 
   const checkHeight = () => {
     if (!path && height >= 500) return 'background';
@@ -29,17 +23,17 @@ const NavBar = () => {
 
   return (
     <div className={`header ${checkHeight()}`}>
-      <div className="container">
-        <div className="header-content">
-          <div className="logo">
-            <NavLink className="logo-link" to="/">
+      <div className='container'>
+        <div className='header-content'>
+          <div className='logo'>
+            <NavLink className='logo-link' to='/'>
               SpaceX
             </NavLink>
           </div>
           {!path && (
-            <ul className="nav">
+            <ul className='nav'>
               <li>
-                <button className="active" onClick={() => scrollToElement(0)}>
+                <button className='active' onClick={() => scrollToElement(0)}>
                   Home
                 </button>
               </li>
@@ -55,6 +49,4 @@ const NavBar = () => {
       </div>
     </div>
   );
-};
-
-export default NavBar;
+}
