+++ src/components/SingleLaunch/SingleLaunch.jsx
@@ -1,63 +1,53 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useParams } from 'react-router-dom';
 import SingleLaunchLoading from './SingleLaunchLoading';
 import NotFoundLaunches from '../Launches/NotFoundLaunches';
 import './singleLaunch.css';
+import useLaunchHook from '../../hooks/useLaunchHook';
 
 const SingleLaunch = () => {
-  const [launch, setLaunch] = useState({});
-  const [error, setError] = useState(false);
-  const [loading, setLoading] = useState(true);
-
   const { id } = useParams();
+  const [launch, error, loading] = useLaunchHook(id);
 
-  useEffect(() => {
-    const abortController = new AbortController();
-    fetch(`https://api.spacexdata.com/v4/launches/${id}`, {
-      signal: abortController.signal,
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        setLaunch(data);
-        setError(false);
-        setLoading(false);
-      })
-      .catch(() => {
-        setError(true);
-        setLoading(false);
-      });
-
-    return () => abortController.abort();
-  }, [id]);
-
+  /**
+   * check Has Launch
+   */
   const checkLaunch = () => Object.keys(launch).length;
 
+  /**
+   * Loading Component
+   */
   const Loading = () => loading && <SingleLaunchLoading />;
 
+  /**
+   * Error Component
+   */
   const Error = () => error && <NotFoundLaunches />;
 
+  /**
+   * Launch Component
+   */
+  const { name, success, details, links } = launch;
   const Launch = () => (
     <div>
       {checkLaunch() && (
         <>
           <div className='launch-head'>
-            <h2 className='launch-title'>{launch.name}</h2>
-            {launch.success ? (
-              <p className='patch success'>Success</p>
-            ) : (
-              <p className='patch failed'>Failed</p>
-            )}
+            <h2 className='launch-title'>{name}</h2>
+            <p className={`patch ${success ? 'success' : 'failed'}`}>
+              {success ? 'Success' : 'Failed'}
+            </p>
           </div>
-          <p className='launch-desc'>{launch.details}</p>
+          <p className='launch-desc'>{details}</p>
           <button
-            onClick={() => window.open(launch.links.article, '_blank')}
+            onClick={() => window.open(links.article, '_blank')}
             className='read-more'>
             Read More
           </button>
           <iframe
             width='1350'
             height='480'
-            src={`https://www.youtube.com/embed/${launch.links.youtube_id}`}
+            src={`https://www.youtube.com/embed/${links.youtube_id}`}
             title='YouTube video player'
             frameBorder='0'
             allow='accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture'
